requirements.txt
flask
psycopg2-binary
gunicorn
requests
schedule
